# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NovaTeck API",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "register",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Missing, invalid or expired token"
          },
          "403": {
            "description": "Owner or administrator access required"
          },
          "503": {
            "description": "Service unavailable"
          },
          "201": {
            "description": "Account created"
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "skills": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "login",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Missing, invalid or expired token"
          },
          "403": {
            "description": "Owner or administrator access required"
          },
          "503": {
            "description": "Service unavailable"
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "remember": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/password-reset/request": {
      "post": {
        "summary": "Create a one-time reset token without exposing whether an email exists.",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Missing, invalid or expired token"
          },
          "403": {
            "description": "Owner or administrator access required"
          },
          "503": {
            "description": "Service unavailable"
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/password-reset/confirm": {
      "post": {
        "summary": "confirm password reset",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Missing, invalid or expired token"
          },
          "403": {
            "description": "Owner or administrator access required"
          },
          "503": {
            "description": "Service unavailable"
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/jobs": {
      "get": {
        "summary": "get jobs",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Missing, invalid or expired token"
          },
          "403": {
            "description": "Owner or administrator access required"
          },
          "503": {
            "description": "Service unavailable"
          }
        },
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skill",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "location",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "company_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "experience",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/jobs/count": {
      "get": {
        "summary": "Return a filtered total for pagination without transferring every row.",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Missing, invalid or expired token"
          },
          "403": {
            "description": "Owner or administrator access required"
          },
          "503": {
            "description": "Service unavailable"
          }
        },
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skill",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "location",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "company_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "experience",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/search-options": {
      "get": {
        "summary": "search options",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Missing, invalid or expired token"
          },
          "403": {
            "description": "Owner or administrator access required"
          },
          "503": {
            "description": "Service unavailable"
          }
        }
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "summary": "get job",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Missing, invalid or expired token"
          },
          "403": {
            "description": "Owner or administrator access required"
          },
          "503": {
            "description": "Service unavailable"
          }
        },
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
    "/saved-jobs": {
      "post": {
        "summary": "save job",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Missing, invalid or expired token"
          },
          "403": {
            "description": "Owner or administrator access required"
          },
          "503": {
            "description": "Service unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "job_id": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/saved-jobs/{job_id}": {
      "delete": {
        "summary": "remove saved job",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Missing, invalid or expired token"
          },
          "403": {
            "description": "Owner or administrator access required"
          },
          "503": {
            "description": "Service unavailable"
          }
        },
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
...
    "/admin/flagged-skills": {
      "get": {
        "summary": "List pending skills for authorized administrators.",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Missing, invalid or expired token"
          },
          "403": {
            "description": "Owner or administrator access required"
          },
          "503": {
            "description": "Service unavailable"
          },
          "409": {
            "description": "Flag missing or already reviewed"
          }
        }
      }
    },
...
  }
}